setup the home component
